Changed telegram api library.  Added some repository and service methods. Added telegram bot class.

+++ src/main/java/com/ness/telegram/wishlistbot/repository/WishRepository.java
@@ -11,6 +11,6 @@ import org.springframework.stereotype.Repository;
 @Transactional
 public interface WishRepository extends JpaRepository<Wish, Long>{
     Wish findByLabelAndUser(String label, User user);
-    Wish findByLabelAndUserChatId(String label, Integer chatId);
-    List<Wish> findByUserChatId(Integer chatId);
+    Wish findByLabelAndUserChatId(String label, Long chatId);
+    List<Wish> findByUserChatId(Long chatId);
 }
